hoxfix: make the test suite more robust
The old test suite used a lot of snapshots to check the result matrix. This is problematic. First, it cannot assure the correctness of the result. Second, it leads to R CMD check fail with Github Actions due to the floating precision problem. We have now upgrade the whole test suite to use math property-based tests.